request-queue/requirements.txt
fastapi
uvicorn
google-cloud-pubsub
google-cloud-storage
google-cloud-logging
pydantic
redis 